# Make validating addresses in spreadsheets about 5× faster 