--- a/notifications_utils/postal_address.py
+++ b/notifications_utils/postal_address.py
@@ -23,6 +23,7 @@ address_lines_1_to_6_keys = [
 address_lines_1_to_6_and_postcode_keys = address_lines_1_to_6_keys + ['postcode']
 address_line_7_key = 'address_line_7'
 address_lines_1_to_7_keys = address_lines_1_to_6_keys + [address_line_7_key]
+country_UK = Country(UK)
 
 
 class PostalAddress():
@@ -46,7 +47,7 @@ class PostalAddress():
             self._lines_without_country = self._lines[:-1]
         except CountryNotFoundError:
             self._lines_without_country = self._lines
-            self.country = Country(UK)
+            self.country = country_UK
 
     def __bool__(self):
         return bool(self.normalised)