# Summary 

We’ve noticed some users getting `504` errors when uploading big CSV files. I think that this problem is specific to letters. Processing a file with lots of rows of address data has got a lot slower since we started validating each row as an address<sup>1</sup>.

The request timeout on our app servers is 60 seconds, so if it takes longer than that to process the CSV the user will see an error.

This pull request makes a series of optimisations to make the validation of addresses about 5× faster. 

Our aim should be to return a response in under 10 seconds, because this about the length of time a user will wait before going off and doing something else<sup>2</sup>. These optimisations get us in the region of 10 seconds, depending on machine speed and how much else is going on in the request.

# Before and after

Validating a CSV file with 50,000 unique but correct UK addresses. Run on a mid 2015 MacBook Pro with a 2.2 GHz Intel Core i7.

Profiler visualisations generated with [Snakeviz](https://jiffyclub.github.io/snakeviz/).

## Before 

66 seconds to run and profile the test. Almost all of the time is spent in the `PostalAddress` code:

<img width="1017" alt="Screenshot 2020-06-13 at 17 16 54" src="https://user-images.githubusercontent.com/355079/84574498-b61ccc00-ad9e-11ea-86b7-93020e194a6d.png">

## After 

12 seconds to run and profile the test. About half of the time of the time is spent in the `PostalAddress` code:

<img width="1022" alt="Screenshot 2020-06-13 at 17 15 05" src="https://user-images.githubusercontent.com/355079/84574508-c2088e00-ad9e-11ea-8e0d-26637f0c81b5.png">

# How each commit contributes to the improvement

Most of the time saved comes from not re-running the same code that normalises the address. But there are some other smaller optimisations that give us another 30% improvement.

These timings are faster than those given above because they’re run without profiling.

Commit | Time to run test | Improvement
-------|------------------|------------
Add test for processing a big CSV of addresses | 38.11s | 0.00×
Compute frequently used properties in constructor of `PostalAddress` | 11.38s | 3.34×
Put LRU cache on postcode detection and formatting | 9.81s | 3.88×
Only compute Country instance for UK once | 9.64s | 3.95×
Optimise country lookup when given a postcode | 8.27s | 4.61×
Reduce number of lines in big address CSV test | 0.91s | N/A

***

1. We used to check only that there was content in 3 of the address columns – see https://github.com/alphagov/notifications-utils/pull/724
2. https://www.nngroup.com/articles/response-times-3-important-limits/